+++ starsky/starsky/clientapp/src/contexts/archive-context.spec.tsx
@@ -221,6 +221,25 @@ describe("ArchiveContext", () => {
     expect(result.fileIndexItems[0].filePath).toBe("/test1.jpg");
   });
 
+  it("remove - last colorClassUsage is removed", () => {
+    var state = {
+      fileIndexItems: [
+        {
+          filePath: "/test.jpg"
+        }
+      ],
+      colorClassUsage: [1, 2]
+    } as IArchiveProps;
+
+    // fullpath input
+    var action = { type: "remove", toRemoveFileList: ["/test.jpg"] } as any;
+
+    var result = archiveReducer(state, action);
+
+    expect(result.fileIndexItems.length).toBe(0);
+    expect(result.colorClassUsage).toStrictEqual([]);
+  });
+
   it("update - check if item is update (append false)", () => {
     var state = {
       ...newIArchive(),
@@ -437,6 +456,91 @@ describe("ArchiveContext", () => {
       {
         fileName: "test3.jpg",
         filePath: "/test3.jpg",
+        status: IExifStatus.Ok
+      }
+    ]);
+  });
+
+  it("add -- remove a collection item", () => {
+    // current state
+    var state = {
+      fileIndexItems: [
+        {
+          fileName: "test0.jpg",
+          filePath: "/test0.jpg",
+          status: IExifStatus.Ok
+        },
+        {
+          fileName: "test0.mp4",
+          filePath: "/test0.mp4",
+          status: IExifStatus.Ok
+        }
+      ],
+      collections: true
+    } as IArchiveProps;
+
+    // to add/remove this
+    var add = [
+      {
+        fileName: "test0.mp4",
+        filePath: "/test0.mp4",
+        status: IExifStatus.Deleted
+      }
+    ];
+    var action = { type: "add", add } as any;
+    var result = archiveReducer(state, action);
+
+    expect(result.fileIndexItems.length).toBe(1);
+    expect(result.fileIndexItems).toStrictEqual([
+      {
+        fileName: "test0.jpg",
+        filePath: "/test0.jpg",
+        status: IExifStatus.Ok
+      }
+    ]);
+  });
+
+  it("add -- implicit delete", () => {
+    console.log("-implicit delete");
+
+    // current state
+    var state = {
+      fileIndexItems: [
+        {
+          fileName: "test0.jpg",
+          fileCollectionName: "test0",
+          filePath: "/test0.jpg",
+          status: IExifStatus.Ok
+        },
+        {
+          fileName: "test1.jpg",
+          fileCollectionName: "test1",
+          filePath: "/test1.jpg",
+          status: IExifStatus.Ok
+        }
+      ],
+      collections: true
+    } as IArchiveProps;
+
+    // to add/remove this
+    var add = [
+      {
+        fileName: "test0.jpg",
+        filePath: "/test0.jpg",
+        status: IExifStatus.Deleted
+      }
+    ];
+    var action = { type: "add", add } as any;
+    var result = archiveReducer(state, action);
+
+    console.log("<--");
+
+    expect(result.fileIndexItems.length).toBe(1);
+    expect(result.fileIndexItems).toStrictEqual([
+      {
+        fileName: "test1.jpg",
+        fileCollectionName: "test1",
+        filePath: "/test1.jpg",
         status: IExifStatus.Ok
       }
     ]);